PR(CHORE): Organize dac and nac into their appropriate files
